// src/app/models/product-calculation.model.ts
export interface ProductCalculation {
  feet: number;
  inch: number;
  nos: number;
  runningFeet: number;
  sqFeet: number;
  weight: number;
}

export interface ProductCalculationTotal {
  totalFeet: number;
  totalInch: number;
  totalNos: number;
  totalRunningFeet: number;
  totalSqFeet: number;
  totalWeight: number;
} 